Remove unused lmdb code
