+++ keriox_core/src/database/mod.rs
@@ -1,250 +1,481 @@
+pub(crate) mod tables;
+
+use std::path::{Path, PathBuf};
+
+use self::tables::{SledEventTree, SledEventTreeVec};
+#[cfg(feature = "query")]
+use crate::query::reply_event::SignedReply;
 use crate::{
-    event_parsing::message::message,
-    prefix::{
-        AttachedSignaturePrefix, BasicPrefix, IdentifierPrefix, SelfAddressingPrefix,
-        SelfSigningPrefix,
+    event::EventMessage,
+    event_message::{
+        key_event_message::KeyEvent,
+        signed_event_message::{
+            SignedEventMessage, SignedNontransferableReceipt, SignedTransferableReceipt,
+            TimestampedNontransReceipt, TimestampedSignedEventMessage, TimestampedTransReceipt,
+        },
+        TimestampedEventMessage,
     },
-    state::IdentifierState,
+    prefix::IdentifierPrefix,
 };
-#[cfg(feature = "lmdb")]
-pub mod lmdb;
 
-#[cfg(feature = "sled-db")]
-pub mod sled;
-/// Event Database
-///
-/// An Abstract model of state for Key Events,
-/// Signatures and Receipts.
-pub trait EventDatabase {
-    type Error;
+pub struct SledEventDatabase {
+    // "iids" tree
+    // this thing is expensive, but everything else is cheeeeeep
+    identifiers: SledEventTree<IdentifierPrefix>,
+    // "kels" tree
+    key_event_logs: SledEventTreeVec<TimestampedSignedEventMessage>,
+    // "ooes" tree
+    escrowed_out_of_order: SledEventTreeVec<TimestampedSignedEventMessage>,
+    // "pses" tree
+    escrowed_partially_signed: SledEventTreeVec<TimestampedSignedEventMessage>,
+    // "ldes" tree
+    likely_duplicious_events: SledEventTreeVec<TimestampedEventMessage>,
+    // "dels" tree
+    duplicitous_events: SledEventTreeVec<TimestampedSignedEventMessage>,
+    // "rcts" tree
+    receipts_nt: SledEventTreeVec<SignedNontransferableReceipt>,
+    // "ures" tree
+    escrowed_receipts_nt: SledEventTreeVec<TimestampedNontransReceipt>,
+    // "vrcs" tree
+    receipts_t: SledEventTreeVec<SignedTransferableReceipt>,
+    // "vres" tree
+    escrowed_receipts_t: SledEventTreeVec<TimestampedTransReceipt>,
+    // "pwes" tree
+    partially_witnessed_events: SledEventTreeVec<TimestampedSignedEventMessage>,
 
-    /// Last Event At SN
-    ///
-    /// Returns the raw bytes of the last event inserted
-    /// for a given identifier at a given sequence number
-    /// (rotation events can supercede interaction events
-    /// with the same sn)
-    /// TODO: see if it can return a Result<Option<&'a [u8]>_>
-    fn last_event_at_sn(
-        &self,
-        pref: &IdentifierPrefix,
-        sn: u64,
-    ) -> Result<Option<Vec<u8>>, Self::Error>;
+    #[cfg(feature = "query")]
+    accepted_rpy: SledEventTreeVec<SignedReply>,
 
-    /// Get State for Prefix
-    ///
-    /// Returns the current State associated with
-    /// the given Prefix
-    fn get_state_for_prefix(
-        &self,
-        pref: &IdentifierPrefix,
-    ) -> Result<Option<IdentifierState>, Self::Error> {
-        // start with empty state
-        let mut state = IdentifierState::default();
+    #[cfg(feature = "query")]
+    escrowed_replys: SledEventTreeVec<SignedReply>,
 
-        // starting from inception
-        for sn in 0.. {
-            // read the latest raw event
-            let raw = match self.last_event_at_sn(pref, sn)? {
-                Some(r) => r,
-                None => {
-                    if sn == 0 {
-                        // no inception event, no state
-                        return Ok(None);
-                    } else {
-                        // end of KEL, stop looping
-                        break;
-                    }
-                }
-            };
-            // parse event
-            // FIXME, DONT UNWRAP
-            let parsed = message(&raw).unwrap().1;
-            // apply it to the state
-            // TODO avoid .clone()
-            state = match state.clone().apply(&parsed) {
-                Ok(s) => s,
-                // will happen when a recovery has overridden some part of the KEL,
-                // stop processing here
-                Err(_) => break,
-            }
-        }
-
-        Ok(Some(state))
-    }
-
-    /// Get KERL for Prefix
-    ///
-    /// Returns the current validated KERL for a given Prefix
-    fn get_kerl(&self, id: &IdentifierPrefix) -> Result<Option<Vec<u8>>, Self::Error>;
-
-    /// Get Children of Prefix
-    ///
-    /// Returns the Identifiers delegated to by the
-    /// given Prefix
-    fn get_children_of_prefix(
-        &self,
-        _pref: &IdentifierPrefix,
-    ) -> Result<Option<Vec<IdentifierPrefix>>, Self::Error> {
-        todo!()
-    }
-
-    /// Get Parent of Prefix
-    ///
-    /// Returns the delegator for the given Prefix,
-    /// if there is one
-    fn get_parent_of_prefix(
-        &self,
-        _pref: &IdentifierPrefix,
-    ) -> Result<Option<IdentifierPrefix>, Self::Error> {
-        todo!()
-    }
-
-    /// Log Event
-    ///
-    /// Adds the raw event data to the database and a timestamp
-    fn log_event(
-        &self,
-        prefix: &IdentifierPrefix,
-        dig: &SelfAddressingPrefix,
-        raw: &[u8],
-        sigs: &[AttachedSignaturePrefix],
-    ) -> Result<(), Self::Error>;
-
-    /// Finalise Event
-    ///
-    /// Update associated logs for fully verified event
-    fn finalise_event(
-        &self,
-        prefix: &IdentifierPrefix,
-        sn: u64,
-        dig: &SelfAddressingPrefix,
-    ) -> Result<(), Self::Error>;
-
-    /// Escrow Partially Signed Event
-    ///
-    /// Escrows an Event which does not have enough signatures
-    fn escrow_partially_signed_event(
-        &self,
-        pref: &IdentifierPrefix,
-        sn: u64,
-        dig: &SelfAddressingPrefix,
-    ) -> Result<(), Self::Error>;
-
-    /// Escrow Out of Order Event
-    ///
-    /// Escrows an Event which has arrived before previous events
-    fn escrow_out_of_order_event(
-        &self,
-        pref: &IdentifierPrefix,
-        sn: u64,
-        dig: &SelfAddressingPrefix,
-    ) -> Result<(), Self::Error>;
-
-    /// Likely Duplicitous Event
-    ///
-    /// Marks an event as being likely duplicitous
-    fn likely_duplicitous_event(
-        &self,
-        pref: &IdentifierPrefix,
-        sn: u64,
-        dig: &SelfAddressingPrefix,
-    ) -> Result<(), Self::Error>;
-
-    /// Duplicitous Event
-    ///
-    /// Marks an event as being known duplicitous
-    fn duplicitous_event(
-        &self,
-        pref: &IdentifierPrefix,
-        sn: u64,
-        dig: &SelfAddressingPrefix,
-    ) -> Result<(), Self::Error>;
-
-    /// Add Non-Transferrable Receipt
-    ///
-    /// Associates a signature Sig made by Signer with the event referenced by Dig and Pref
-    fn add_nt_receipt_for_event(
-        &self,
-        pref: &IdentifierPrefix,
-        dig: &SelfAddressingPrefix,
-        signer: &BasicPrefix,
-        sig: &SelfSigningPrefix,
-    ) -> Result<(), Self::Error>;
-
-    /// Add Transferrable Receipt
-    ///
-    /// Associates a signature Sig made by Signer with the event referenced by Dig and Pref
-    fn add_t_receipt_for_event(
-        &self,
-        pref: &IdentifierPrefix,
-        dig: &SelfAddressingPrefix,
-        signer: &IdentifierPrefix,
-        sig: &AttachedSignaturePrefix,
-    ) -> Result<(), Self::Error>;
-
-    /// Escrow Non-Transferrable Receipt
-    fn escrow_nt_receipt(
-        &self,
-        pref: &IdentifierPrefix,
-        dig: &SelfAddressingPrefix,
-        signer: &BasicPrefix,
-        sig: &SelfSigningPrefix,
-    ) -> Result<(), Self::Error>;
-
-    /// Escrow Transferrable Receipt
-    fn escrow_t_receipt(
-        &self,
-        pref: &IdentifierPrefix,
-        dig: &SelfAddressingPrefix,
-        signer: &IdentifierPrefix,
-        sig: &AttachedSignaturePrefix,
-    ) -> Result<(), Self::Error>;
-
-    fn has_receipt(
-        &self,
-        pref: &IdentifierPrefix,
-        sn: u64,
-        validator: &IdentifierPrefix,
-    ) -> Result<bool, Self::Error>;
+    #[cfg(feature = "query")]
+    mailbox_receipts: SledEventTreeVec<SignedNontransferableReceipt>,
 }
 
-// #[cfg(test)]
-// fn test_db<D: EventDatabase>(db: D) -> Result<(), D::Error> {
-//     use crate::{
-//         derivation::self_addressing::SelfAddressing,
-//         event_message::parse::{signed_message, Deserialized},
-//     };
+// TODO: remove all the `.ok()`s
+impl SledEventDatabase {
+    pub fn new<'a, P>(path: P) -> Result<Self, DbError>
+    where
+        P: Into<&'a Path>,
+    {
+        let mut events_path = PathBuf::new();
+        events_path.push(&path.into());
+        let mut escrow_path = events_path.clone();
 
-//     // NOTE this is not actually a valid kel (sig indexes and prefix are wrong), just has the correct form
-//     let signed = br#"{"v":"KERI10JSON000144_","i":"E005TfcIFvrzhJFxoGqebPHvtanxEwcfJOAYcUtCmhk8","s":"0","t":"icp","kt":"1","k":["D5UzOMC5Knhi5eA-Cr8ASuD8lUcZMcLtAhIZ33W5Z4hs","DEbposribdTgsnCSQgmVN6VKoc4Vpc-hs9rbskXQ2O2M","D46n6npISQETk7eGYnwe5Jq7USmEsckHeJRu2YoTCXhU"],"n":"E2zBfVYkE2uaGR5DmMVBbWsIdBIZVu5Ml6joenraD5Ho","bt":"0","b":[],"c":[],"a":[]}-AADAAKfgMIEsKlrXqUxUyw1Qq7gFrg9mNWcDAkQAXUW6Hppvt4NBdEbU_2Wy7Re0zp5zVvLjjq4hjPE5aXSeIlHh7DgABIkKK8jI0l_tor7EaIM2B65aLn9e6Y3Igwa9OjDqbiyqXdL1yHxga7nhJY80Ct0zXGhm7hgLzgB6d86EqfXWLCQACBPCeqK_WXY64EZ8E91Y2trI6MfZT-f2NmtHCmmKhvt7AmehPcvQSvrcQbogdNEBr749AbVG7glVsV8WitVR2DQ"#;
-//     let deser = match signed_message(signed).unwrap().1 {
-//         Deserialized::Event(e) => e,
-//         _ => panic!(),
-//     };
+        events_path.push("events");
+        escrow_path.push("escrow");
 
-//     let dig = SelfAddressing::Blake3_256.derive(deser.event.raw);
+        let db = sled::open(events_path.as_path())?;
+        let escrows_db = sled::open(escrow_path.as_path())?;
 
-//     db.log_event(
-//         &deser.event.event.event.prefix,
-//         &dig,
-//         deser.event.raw,
-//         &deser.signatures,
-//     )?;
+        Ok(Self {
+            identifiers: SledEventTree::new(db.open_tree(b"iids")?),
+            receipts_t: SledEventTreeVec::new(db.open_tree(b"vrcs")?),
+            receipts_nt: SledEventTreeVec::new(db.open_tree(b"rcts")?),
+            key_event_logs: SledEventTreeVec::new(db.open_tree(b"kels")?),
+            likely_duplicious_events: SledEventTreeVec::new(db.open_tree(b"ldes")?),
+            duplicitous_events: SledEventTreeVec::new(db.open_tree(b"dels")?),
+            #[cfg(feature = "query")]
+            accepted_rpy: SledEventTreeVec::new(db.open_tree(b"knas")?),
+            #[cfg(feature = "query")]
+            mailbox_receipts: SledEventTreeVec::new(db.open_tree(b"mbxr")?),
 
-//     db.finalise_event(&deser.event.event.event.prefix, 0, &dig)?;
+            escrowed_out_of_order: SledEventTreeVec::new(escrows_db.open_tree(b"ooes")?),
+            escrowed_partially_signed: SledEventTreeVec::new(escrows_db.open_tree(b"pses")?),
+            partially_witnessed_events: SledEventTreeVec::new(escrows_db.open_tree(b"pwes")?),
+            escrowed_receipts_nt: SledEventTreeVec::new(escrows_db.open_tree(b"ures")?),
+            escrowed_receipts_t: SledEventTreeVec::new(escrows_db.open_tree(b"vres")?),
+            #[cfg(feature = "query")]
+            escrowed_replys: SledEventTreeVec::new(escrows_db.open_tree(b"knes")?),
+        })
+    }
 
-//     let written = db.last_event_at_sn(&deser.event.event.event.prefix, 0)?;
+    pub fn add_kel_finalized_event(
+        &self,
+        event: SignedEventMessage,
+        id: &IdentifierPrefix,
+    ) -> Result<(), DbError> {
+        self.key_event_logs
+            .push(self.identifiers.designated_key(id)?, event.into())
+    }
 
-//     assert_eq!(written, Some(deser.event.raw.to_vec()));
+    pub fn get_kel_finalized_events(
+        &self,
+        id: &IdentifierPrefix,
+    ) -> Option<impl DoubleEndedIterator<Item = TimestampedSignedEventMessage>> {
+        self.key_event_logs
+            .iter_values(self.identifiers.designated_key(id).ok()?)
+    }
 
-//     let kerl = db.get_kerl(&deser.event.event.event.prefix)?.unwrap();
-//     let deser_2 = match signed_message(&kerl).unwrap().1 {
-//         Deserialized::Event(e) => e,
-//         _ => panic!(),
-//     };
+    pub fn remove_kel_finalized_event(
+        &self,
+        id: &IdentifierPrefix,
+        event: &SignedEventMessage,
+    ) -> Result<(), DbError> {
+        self.key_event_logs
+            .remove(self.identifiers.designated_key(id)?, &event.into())
+    }
 
-//     assert_eq!(deser, deser_2);
+    pub fn add_out_of_order_event(
+        &self,
+        event: SignedEventMessage,
+        id: &IdentifierPrefix,
+    ) -> Result<(), DbError> {
+        self.escrowed_out_of_order
+            .push(self.identifiers.designated_key(id)?, event.into())
+    }
 
-//     Ok(())
-// }
+    pub fn get_out_of_order_events(
+        &self,
+        id: &IdentifierPrefix,
+    ) -> Option<impl DoubleEndedIterator<Item = TimestampedSignedEventMessage>> {
+        self.escrowed_out_of_order
+            .iter_values(self.identifiers.designated_key(id).ok()?)
+    }
+
+    pub fn remove_out_of_order_event(
+        &self,
+        id: &IdentifierPrefix,
+        event: &SignedEventMessage,
+    ) -> Result<(), DbError> {
+        self.escrowed_out_of_order
+            .remove(self.identifiers.designated_key(id)?, &event.into())
+    }
+
+    pub fn add_partially_signed_event(
+        &self,
+        event: SignedEventMessage,
+        id: &IdentifierPrefix,
+    ) -> Result<(), DbError> {
+        self.escrowed_partially_signed
+            .push(self.identifiers.designated_key(id)?, event.into())
+    }
+
+    pub fn get_all_partially_signed_events(
+        &self,
+    ) -> Option<impl DoubleEndedIterator<Item = TimestampedSignedEventMessage>> {
+        self.escrowed_partially_signed.get_all()
+    }
+
+    pub fn get_partially_signed_events(
+        &self,
+        event: EventMessage<KeyEvent>,
+    ) -> Option<impl DoubleEndedIterator<Item = TimestampedSignedEventMessage>> {
+        self.escrowed_partially_signed
+            .iter_values(
+                self.identifiers
+                    .designated_key(&event.event.get_prefix())
+                    .ok()?,
+            )
+            .map(|events| {
+                events
+                    .filter(move |db_event| event.eq(&db_event.signed_event_message.event_message))
+            })
+    }
+
+    pub fn remove_partially_signed_event(
+        &self,
+        id: &IdentifierPrefix,
+        event: &EventMessage<KeyEvent>,
+    ) -> Result<(), DbError> {
+        if let Some(partially_signed) = self.get_partially_signed_events(event.clone()) {
+            for partially_event in partially_signed {
+                self.escrowed_partially_signed
+                    .remove(self.identifiers.designated_key(id)?, &partially_event)?;
+            }
+        }
+        Ok(())
+    }
+
+    pub fn add_partially_witnessed_event(
+        &self,
+        event: SignedEventMessage,
+        id: &IdentifierPrefix,
+    ) -> Result<(), DbError> {
+        let event = event.into();
+        if !self.partially_witnessed_events.contains_value(&event) {
+            self.partially_witnessed_events
+                .push(self.identifiers.designated_key(id)?, event)
+        } else {
+            Ok(())
+        }
+    }
+
+    pub fn get_partially_witnessed_events(
+        &self,
+        id: &IdentifierPrefix,
+    ) -> Option<impl DoubleEndedIterator<Item = TimestampedSignedEventMessage>> {
+        self.partially_witnessed_events
+            .iter_values(self.identifiers.designated_key(id).ok()?)
+    }
+
+    pub fn remove_partially_witnessed_event(
+        &self,
+        id: &IdentifierPrefix,
+        event: &SignedEventMessage,
+    ) -> Result<(), DbError> {
+        self.partially_witnessed_events
+            .remove(self.identifiers.designated_key(id)?, &event.into())
+    }
+
+    pub fn get_all_partially_witnessed(
+        &self,
+    ) -> Option<impl DoubleEndedIterator<Item = TimestampedSignedEventMessage>> {
+        self.partially_witnessed_events.get_all()
+    }
+
+    pub fn add_receipt_t(
+        &self,
+        receipt: SignedTransferableReceipt,
+        id: &IdentifierPrefix,
+    ) -> Result<(), DbError> {
+        self.receipts_t
+            .push(self.identifiers.designated_key(id)?, receipt)
+    }
+
+    pub fn get_receipts_t(
+        &self,
+        id: &IdentifierPrefix,
+    ) -> Option<impl DoubleEndedIterator<Item = SignedTransferableReceipt>> {
+        self.receipts_t
+            .iter_values(self.identifiers.designated_key(id).ok()?)
+    }
+
+    pub fn add_receipt_nt(
+        &self,
+        receipt: SignedNontransferableReceipt,
+        id: &IdentifierPrefix,
+    ) -> Result<(), DbError> {
+        self.receipts_nt
+            .push(self.identifiers.designated_key(id)?, receipt)
+    }
+
+    pub fn get_receipts_nt(
+        &self,
+        id: &IdentifierPrefix,
+    ) -> Option<impl DoubleEndedIterator<Item = SignedNontransferableReceipt>> {
+        self.receipts_nt
+            .iter_values(self.identifiers.designated_key(id).ok()?)
+    }
+
+    pub fn remove_receipts_nt(&self, id: &IdentifierPrefix) -> Result<(), DbError> {
+        if let Some(receipts) = self.get_receipts_nt(id) {
+            for receipt in receipts {
+                self.receipts_nt
+                    .remove(self.identifiers.designated_key(id)?, &receipt)?;
+            }
+        }
+        Ok(())
+    }
+
+    pub fn add_escrow_t_receipt(
+        &self,
+        receipt: SignedTransferableReceipt,
+        id: &IdentifierPrefix,
+    ) -> Result<(), DbError> {
+        self.escrowed_receipts_t
+            .push(self.identifiers.designated_key(id)?, receipt.into())
+    }
+
+    pub fn get_escrow_t_receipts(
+        &self,
+        id: &IdentifierPrefix,
+    ) -> Option<impl DoubleEndedIterator<Item = TimestampedTransReceipt>> {
+        self.escrowed_receipts_t
+            .iter_values(self.identifiers.designated_key(id).ok()?)
+    }
+
+    pub fn remove_escrow_t_receipt(
+        &self,
+        id: &IdentifierPrefix,
+        receipt: &SignedTransferableReceipt,
+    ) -> Result<(), DbError> {
+        self.escrowed_receipts_t
+            .remove(self.identifiers.designated_key(id)?, &receipt.into())
+    }
+
+    pub fn add_escrow_nt_receipt(
+        &self,
+        receipt: SignedNontransferableReceipt,
+        id: &IdentifierPrefix,
+    ) -> Result<(), DbError> {
+        let receipt: TimestampedNontransReceipt = receipt.into();
+        if !self.escrowed_receipts_nt.contains_value(&receipt) {
+            self.escrowed_receipts_nt
+                .push(self.identifiers.designated_key(id)?, receipt)
+        } else {
+            Ok(())
+        }
+    }
+
+    pub fn get_escrow_nt_receipts(
+        &self,
+        id: &IdentifierPrefix,
+    ) -> Option<impl DoubleEndedIterator<Item = TimestampedNontransReceipt>> {
+        self.escrowed_receipts_nt
+            .iter_values(self.identifiers.designated_key(id).ok()?)
+    }
+
+    pub fn get_all_escrow_nt_receipts(
+        &self,
+    ) -> Option<impl DoubleEndedIterator<Item = TimestampedNontransReceipt>> {
+        self.escrowed_receipts_nt.get_all()
+    }
+
+    pub fn remove_escrow_nt_receipt(
+        &self,
+        id: &IdentifierPrefix,
+        receipt: &SignedNontransferableReceipt,
+    ) -> Result<(), DbError> {
+        self.escrowed_receipts_nt
+            .remove(self.identifiers.designated_key(id)?, &receipt.into())
+    }
+
+    pub fn add_likely_duplicious_event(
+        &self,
+        event: EventMessage<KeyEvent>,
+        id: &IdentifierPrefix,
+    ) -> Result<(), DbError> {
+        self.likely_duplicious_events
+            .push(self.identifiers.designated_key(id)?, event.into())
+    }
+
+    pub fn get_likely_duplicitous_events(
+        &self,
+        id: &IdentifierPrefix,
+    ) -> Option<impl DoubleEndedIterator<Item = TimestampedEventMessage>> {
+        self.likely_duplicious_events
+            .iter_values(self.identifiers.designated_key(id).ok()?)
+    }
+
+    pub fn add_duplicious_event(
+        &self,
+        event: SignedEventMessage,
+        id: &IdentifierPrefix,
+    ) -> Result<(), DbError> {
+        self.duplicitous_events
+            .push(self.identifiers.designated_key(id)?, event.into())
+    }
+
+    pub fn get_duplicious_events(
+        &self,
+        id: &IdentifierPrefix,
+    ) -> Option<impl DoubleEndedIterator<Item = TimestampedSignedEventMessage>> {
+        self.duplicitous_events
+            .iter_values(self.identifiers.designated_key(id).ok()?)
+    }
+
+    #[cfg(feature = "query")]
+    pub fn update_accepted_reply(
+        &self,
+        rpy: SignedReply,
+        id: &IdentifierPrefix,
+    ) -> Result<(), DbError> {
+        use crate::query::reply_event::ReplyRoute;
+
+        match self
+            .accepted_rpy
+            .iter_values(self.identifiers.designated_key(id)?)
+        {
+            Some(rpys) => {
+                let filtered = rpys
+                    .filter(|s| match (s.reply.get_route(), rpy.reply.get_route()) {
+                        (ReplyRoute::Ksn(id1, _), ReplyRoute::Ksn(id2, _)) => id1 != id2,
+                        _ => true,
+                    })
+                    .chain(Some(rpy.clone()).into_iter())
+                    .collect();
+                self.accepted_rpy
+                    .put(self.identifiers.designated_key(id)?, filtered)
+            }
+            None => self
+                .accepted_rpy
+                .push(self.identifiers.designated_key(id)?, rpy),
+        }
+    }
+
+    #[cfg(feature = "query")]
+    pub fn get_accepted_replys(
+        &self,
+        id: &IdentifierPrefix,
+    ) -> Option<impl DoubleEndedIterator<Item = SignedReply>> {
+        self.accepted_rpy
+            .iter_values(self.identifiers.designated_key(id).ok()?)
+    }
+
+    #[cfg(feature = "query")]
+    pub fn remove_accepted_reply(
+        &self,
+        id: &IdentifierPrefix,
+        rpy: SignedReply,
+    ) -> Result<(), DbError> {
+        self.accepted_rpy
+            .remove(self.identifiers.designated_key(id)?, &rpy)
+    }
+
+    #[cfg(feature = "query")]
+    pub fn add_escrowed_reply(
+        &self,
+        rpy: SignedReply,
+        id: &IdentifierPrefix,
+    ) -> Result<(), DbError> {
+        self.escrowed_replys
+            .push(self.identifiers.designated_key(id)?, rpy)
+    }
+
+    #[cfg(feature = "query")]
+    pub fn get_escrowed_replys(
+        &self,
+        id: &IdentifierPrefix,
+    ) -> Option<impl DoubleEndedIterator<Item = SignedReply>> {
+        self.escrowed_replys
+            .iter_values(self.identifiers.designated_key(id).ok()?)
+    }
+
+    #[cfg(feature = "query")]
+    pub fn remove_escrowed_reply(
+        &self,
+        id: &IdentifierPrefix,
+        rpy: &SignedReply,
+    ) -> Result<(), DbError> {
+        self.escrowed_replys
+            .remove(self.identifiers.designated_key(id)?, rpy)
+    }
+
+    #[cfg(feature = "query")]
+    pub fn get_all_escrowed_replys(&self) -> Option<impl DoubleEndedIterator<Item = SignedReply>> {
+        self.escrowed_replys.get_all()
+    }
+
+    #[cfg(feature = "query")]
+    pub fn add_mailbox_receipt(
+        &self,
+        receipt: SignedNontransferableReceipt,
+        id: &IdentifierPrefix,
+    ) -> Result<(), DbError> {
+        if !self.mailbox_receipts.contains_value(&receipt) {
+            self.mailbox_receipts
+                .push(self.identifiers.designated_key(id)?, receipt)
+        } else {
+            Ok(())
+        }
+    }
+
+    #[cfg(feature = "query")]
+    pub fn get_mailbox_receipts(
+        &self,
+        id: &IdentifierPrefix,
+    ) -> Option<impl DoubleEndedIterator<Item = SignedNontransferableReceipt>> {
+        self.mailbox_receipts
+            .iter_values(self.identifiers.designated_key(id).ok()?)
+    }
+}
+
+#[derive(Debug, thiserror::Error)]
+pub enum DbError {
+    #[error("sled error: {0}")]
+    Sled(#[from] sled::Error),
+    #[error("serde error: {0}")]
+    Serde(#[from] serde_cbor::Error),
+}
